App/task-service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
marshmallow==3.20.1
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
prometheus-client==0.17.1